gpio - major refactoring of Input() API
Because of a handful of issues with the polling thread, I'm moving towards a
simpler API: Input.changes() function returns a triplet of
rising/falling/current values.  So now, the polling thread is always on for an
Input, and always gathers any signal edges.  Additional, this will be easier to test.

May put the previous wait/call APIs back on top of this when I'm done.