accel - working ioctl() based i2c read/write with repeated restart
Added code for i2c read/write, required by the acceleromter for reading data.
Important note: the i2c_bcm2708 driver has support for this, but it is not
enabled by default.  It can be enabled with:

    sudo sh -c 'echo -n 1 > /sys/module/i2c_bcm2708/parameters/combined'

This either needs to be added to the code, or the install.  If added to the
install, the code should at least check to see that its on, and throw an error
if not.