gpio - permission error fixed
Previously, a permission error was received when trying to set the direction of
a GPIO.  This occured intermittently.  Turns out, the way udev sets up the /sys
entries, the entry can not be ready yet by the time we try to access it.  So
now, we retry accesing the entry for a second or two (doubling the sleep time
of retry each attempt).  The error seems to go away in approx 0.05 seconds
(varies)