accel - workaround for bogus reads from MMA8653
On the MMA8653, reads from an arbitrary register require writing the register
number, a repeated-start, and the reading the register value.

Unfortunately, the MMA8653 appears to be a bit flaky, and although a
repeated-start transaction works most of the time, it sometimes gets
the MMA8653 in a funky state (specifically, it sometimes returns
nonsense values, even though it always correctly responds with ACKs on
the bus).

As a workaround, we do two things:
    - If the register is low, we do not use a repeated-start
      transaction, and instead use os.read() to read from regsiter 0 up
      through the requested registers using the parts auto-increment
      feature.  This works always, but only works for low registers
      (STATUS and OUT_?_?SB)
    - We add a read of one byte after the repeated-start transaction.
      This appears to leave the device in a state where it is
      responding correctly MOST of the time, in testing.  But beware -
      this means you can't trust reads from high registers, and may
      need to retry them.