sound - pop-suppression
Playing sounds doesn't generally require pop-suppression, as the sounds taper
at start or end.  But when sounds are stopped early, or more noticibly, when
Note()s are started or stopped, the start/end is abrupt.  Therefore, we add a
triangular time-window at the start/stop of all sounds, for exactly one audio
buffer.  This is enough to drastically reduce the pop, if not completely
eliminate it.