minecraft - get_heading() now black
get_heading() works by moving the player to a far out platform to get the
heading, then moving the player back to where they were.  Previously the
platform was not enclosed.  Now its fully enclosed.  The effect is that now the
screen simply goes to black briefly when get heading is called - which is less
distrurbing.